+++ worker/src/RTC/Codecs/H264_SVC.cpp
@@ -10,8 +10,7 @@ namespace RTC
 	{
 		/* Class methods. */
 
-		H264_SVC::PayloadDescriptor* H264_SVC::Parse(
-		  const uint8_t* data, size_t len, RTC::RtpPacket::FrameMarking* frameMarking, uint8_t frameMarkingLen)
+		H264_SVC::PayloadDescriptor* H264_SVC::Parse(const uint8_t* data, size_t len)
 		{
 			MS_TRACE();
 
@@ -24,138 +23,96 @@ namespace RTC
 
 			std::unique_ptr<PayloadDescriptor> payloadDescriptor(new PayloadDescriptor());
 
-			// Use frame-marking.
-			if (frameMarking)
+			const uint8_t nal = *data & 0x1F;
+
+			switch (nal)
 			{
-				// Read fields.
-				payloadDescriptor->s       = frameMarking->start;
-				payloadDescriptor->e       = frameMarking->end;
-				payloadDescriptor->i       = frameMarking->independent;
-				payloadDescriptor->d       = frameMarking->discardable;
-				payloadDescriptor->b       = frameMarking->base;
-				payloadDescriptor->tlIndex = frameMarking->tid;
-
-				payloadDescriptor->hasTlIndex = true;
-
-				if (frameMarkingLen >= 2)
+				// Single NAL unit packet.
+				// IDR (instantaneous decoding picture).
+				case 1:
+				case 5:
+				case 7:
+				case 14:
+				case 20:
 				{
-					payloadDescriptor->hasSlIndex = true;
-					payloadDescriptor->slIndex    = frameMarking->lid >> 4 & 0x07;
-				}
+					payloadDescriptor =
+					  H264_SVC::ParseSingleNalu(data, len, std::move(payloadDescriptor), true);
 
-				if (frameMarkingLen == 3)
-				{
-					payloadDescriptor->hasTl0picidx = true;
-					payloadDescriptor->tl0picidx    = frameMarking->tl0picidx;
-				}
-
-				// Detect key frame.
-				if (frameMarking->start && frameMarking->independent)
-				{
-					payloadDescriptor->isKeyFrame = true;
-				}
-			}
-
-			// NOTE: Unfortunately libwebrtc produces wrong Frame-Marking (without i=1
-			// in keyframes) when it uses H264 hardware encoder (at least in Mac):
-			//   https://bugs.chromium.org/p/webrtc/issues/detail?id=10746
-			//
-			// As a temporal workaround, always do payload parsing to detect keyframes
-			// if there is no frame-marking or if there is but keyframe was not
-			// detected above.
-			if (!frameMarking || !payloadDescriptor->isKeyFrame)
-			{
-				const uint8_t nal = *data & 0x1F;
-
-				switch (nal)
-				{
-					// Single NAL unit packet.
-					// IDR (instantaneous decoding picture).
-					case 1:
-					case 5:
-					case 7:
-					case 14:
-					case 20:
+					if (payloadDescriptor == nullptr)
 					{
-						payloadDescriptor =
-						  H264_SVC::ParseSingleNalu(data, len, std::move(payloadDescriptor), true);
+						MS_WARN_DEV("ignoring invalid payload (1)");
+
+						return nullptr;
+					}
+
+					break;
+				}
+
+				// Aggreation packet.
+				// STAP-A.
+				case 24:
+				{
+					size_t offset{ 1 };
+
+					len -= 1;
+
+					// Iterate NAL units.
+					while (len >= 3)
+					{
+						auto naluSize = Utils::Byte::Get2Bytes(data, offset);
+
+						payloadDescriptor = H264_SVC::ParseSingleNalu(
+						  (data + offset + sizeof(naluSize)),
+						  (len - sizeof(naluSize)),
+						  std::move(payloadDescriptor),
+						  true);
 
 						if (payloadDescriptor == nullptr)
 						{
-							MS_WARN_DEV("ignoring invalid payload (1)");
+							MS_WARN_DEV("ignoring invalid payload (2)");
 
 							return nullptr;
 						}
 
-						break;
+						if (payloadDescriptor->isKeyFrame)
+						{
+							break;
+						}
+
+						// Check if there is room for the indicated NAL unit size.
+						if (len < (naluSize + sizeof(naluSize)))
+						{
+							break;
+						}
+
+						offset += naluSize + sizeof(naluSize);
+						len -= naluSize + sizeof(naluSize);
 					}
 
-					// Aggreation packet.
-					// STAP-A.
-					case 24:
+					break;
+				}
+
+				// Aggreation packet.
+				// FU-A, FU-B.
+				case 28:
+				case 29:
+				{
+					const uint8_t startBit = *(data + 1) & 0x80;
+
+					if (startBit == 128)
 					{
-						size_t offset{ 1 };
-
-						len -= 1;
-
-						// Iterate NAL units.
-						while (len >= 3)
-						{
-							auto naluSize = Utils::Byte::Get2Bytes(data, offset);
-
-							payloadDescriptor = H264_SVC::ParseSingleNalu(
-							  (data + offset + sizeof(naluSize)),
-							  (len - sizeof(naluSize)),
-							  std::move(payloadDescriptor),
-							  true);
-
-							if (payloadDescriptor == nullptr)
-							{
-								MS_WARN_DEV("ignoring invalid payload (2)");
-
-								return nullptr;
-							}
-
-							if (payloadDescriptor->isKeyFrame)
-							{
-								break;
-							}
-
-							// Check if there is room for the indicated NAL unit size.
-							if (len < (naluSize + sizeof(naluSize)))
-							{
-								break;
-							}
-
-							offset += naluSize + sizeof(naluSize);
-							len -= naluSize + sizeof(naluSize);
-						}
-
-						break;
+						payloadDescriptor = H264_SVC::ParseSingleNalu(
+						  (data + 1), (len - 1), std::move(payloadDescriptor), (startBit == 128 ? true : false));
 					}
 
-					// Aggreation packet.
-					// FU-A, FU-B.
-					case 28:
-					case 29:
+					if (payloadDescriptor == nullptr)
 					{
-						const uint8_t startBit = *(data + 1) & 0x80;
+						MS_WARN_DEV("ignoring invalid payload (3)");
 
-						if (startBit == 128)
-						{
-							payloadDescriptor = H264_SVC::ParseSingleNalu(
-							  (data + 1), (len - 1), std::move(payloadDescriptor), (startBit == 128 ? true : false));
-						}
-
-						if (payloadDescriptor == nullptr)
-						{
-							MS_WARN_DEV("ignoring invalid payload (3)");
-
-							return nullptr;
-						}
-
-						break;
+						return nullptr;
 					}
+
+					break;
 				}
 			}
 
@@ -246,14 +203,8 @@ namespace RTC
 
 			auto* data = packet->GetPayload();
 			auto len   = packet->GetPayloadLength();
-			RtpPacket::FrameMarking* frameMarking{ nullptr };
-			uint8_t frameMarkingLen{ 0 };
 
-			// Read frame-marking.
-			packet->ReadFrameMarking(&frameMarking, frameMarkingLen);
-
-			PayloadDescriptor* payloadDescriptor =
-			  H264_SVC::Parse(data, len, frameMarking, frameMarkingLen);
+			PayloadDescriptor* payloadDescriptor = H264_SVC::Parse(data, len);
 
 			if (!payloadDescriptor)
 			{
@@ -272,13 +223,6 @@ namespace RTC
 			MS_TRACE();
 
 			MS_DUMP("<H264_SVC::PayloadDescriptor>");
-			MS_DUMP(
-			  "  s:%" PRIu8 "|e:%" PRIu8 "|i:%" PRIu8 "|d:%" PRIu8 "|b:%" PRIu8,
-			  this->s,
-			  this->e,
-			  this->i,
-			  this->d,
-			  this->b);
 			MS_DUMP("  hasSlIndex: %s", this->hasSlIndex ? "true" : "false");
 			MS_DUMP("  hasTlIndex: %s", this->hasTlIndex ? "true" : "false");
 			MS_DUMP("  tl0picidx: %" PRIu8, this->tl0picidx);
@@ -296,7 +240,7 @@ namespace RTC
 		}
 
 		bool H264_SVC::PayloadDescriptorHandler::Process(
-		  RTC::Codecs::EncodingContext* encodingContext, RTC::RtpPacket* /*packet*/, bool& marker)
+		  RTC::Codecs::EncodingContext* encodingContext, RTC::RtpPacket* /*packet*/, bool& /*marker*/)
 		{
 			MS_TRACE();
 
@@ -312,12 +256,7 @@ namespace RTC
 
 			// If packet spatial or temporal layer is higher than maximum announced
 			// one, drop the packet.
-			// clang-format off
-			if (
-				packetSpatialLayer >= context->GetSpatialLayers() ||
-				packetTemporalLayer >= context->GetTemporalLayers()
-			)
-			// clang-format on
+			if (packetSpatialLayer >= context->GetSpatialLayers() || packetTemporalLayer >= context->GetTemporalLayers())
 			{
 				MS_WARN_TAG(
 				  rtp, "too high packet layers %" PRIu8 ":%" PRIu8, packetSpatialLayer, packetTemporalLayer);
@@ -325,7 +264,6 @@ namespace RTC
 				return false;
 			}
 
-			// clang-format off
 			const bool isOldPacket = false;
 
 			// Upgrade current spatial layer if needed.
@@ -352,7 +290,6 @@ namespace RTC
 				if (context->IsKSvc())
 				{
 					if (this->payloadDescriptor->isKeyFrame)
-					// clang-format on
 					{
 						MS_DEBUG_DEV(
 						  "downgrading tmpSpatialLayer from %" PRIu16 " to %" PRIu16 " (packet:%" PRIu8
@@ -369,12 +306,10 @@ namespace RTC
 				// In full SVC we do not need a keyframe.
 				else
 				{
-					// clang-format off
-					if (
-						packetSpatialLayer == context->GetTargetSpatialLayer() &&
-						this->payloadDescriptor->e
-					)
-					// clang-format on
+					// TODO: Here we should check if this is the end of the frame but
+					// that info is only present in frame-marking header (bit e) which is
+					// not implemented (because it's deprecated).
+					if (packetSpatialLayer == context->GetTargetSpatialLayer())
 					{
 						MS_DEBUG_DEV(
 						  "downgrading tmpSpatialLayer from %" PRIu16 " to %" PRIu16 " (packet:%" PRIu8
@@ -402,12 +337,10 @@ namespace RTC
 				// Upgrade current temporal layer if needed.
 				if (context->GetTargetTemporalLayer() > context->GetCurrentTemporalLayer())
 				{
-					// clang-format off
-					if (
-						packetTemporalLayer >= context->GetCurrentTemporalLayer() + 1 &&
-						this->payloadDescriptor->s
-					)
-					// clang-format on
+					// TODO: Here we should check if this is the start of the frame but
+					// that info is only present in frame-marking header (bit s) which is
+					// not implemented (because it's deprecated).
+					if (packetTemporalLayer >= context->GetCurrentTemporalLayer() + 1)
 					{
 						MS_DEBUG_DEV(
 						  "upgrading tmpTemporalLayer from %" PRIu16 " to %" PRIu8 " (packet:%" PRIu8 ":%" PRIu8
@@ -423,12 +356,10 @@ namespace RTC
 				// Downgrade current temporal layer if needed.
 				else if (context->GetTargetTemporalLayer() < context->GetCurrentTemporalLayer())
 				{
-					// clang-format off
-					if (
-						packetTemporalLayer == context->GetTargetTemporalLayer() &&
-						this->payloadDescriptor->e
-					)
-					// clang-format on
+					// TODO: Here we should check if this is the end of the frame but
+					// that info is only present in frame-marking header (bit e) which is
+					// not implemented (because it's deprecated).
+					if (packetTemporalLayer == context->GetTargetTemporalLayer())
 					{
 						MS_DEBUG_DEV(
 						  "downgrading tmpTemporalLayer from %" PRIu16 " to %" PRIu16 " (packet:%" PRIu8
@@ -450,10 +381,13 @@ namespace RTC
 			}
 
 			// Set marker bit if needed.
-			if (packetSpatialLayer == tmpSpatialLayer && this->payloadDescriptor->e)
-			{
-				marker = true;
-			}
+			// TODO: Here we should check if this is the end of the frame but
+			// that info is only present in frame-marking header (bit e) which is
+			// not implemented (because it's deprecated).
+			// if (packetSpatialLayer == tmpSpatialLayer && this->payloadDescriptor->e)
+			// {
+			// 	marker = true;
+			// }
 
 			// Update current spatial layer if needed.
 			if (tmpSpatialLayer != context->GetCurrentSpatialLayer())