+++ worker/src/RTC/RtpDictionaries/RtpCodecMimeType.cpp
@@ -38,7 +38,6 @@ namespace RTC
 		{ "h264",            RtpCodecMimeType::Subtype::H264            },
 		{ "h264-svc",        RtpCodecMimeType::Subtype::H264_SVC        },
 		{ "x-h264uc",        RtpCodecMimeType::Subtype::X_H264UC        },
-		{ "h265",            RtpCodecMimeType::Subtype::H265            },
 		{ "av1",             RtpCodecMimeType::Subtype::AV1             },
 		// Complementary codecs:
 		{ "cn",              RtpCodecMimeType::Subtype::CN              },
@@ -67,7 +66,6 @@ namespace RTC
 		{ RtpCodecMimeType::Subtype::H264,            "H264"            },
 		{ RtpCodecMimeType::Subtype::H264_SVC,        "H264-SVC"        },
 		{ RtpCodecMimeType::Subtype::X_H264UC,        "X-H264UC"        },
-		{ RtpCodecMimeType::Subtype::H265,            "H265"            },
 		{ RtpCodecMimeType::Subtype::AV1,             "AV1"             },
 		// Complementary codecs:
 		{ RtpCodecMimeType::Subtype::CN,              "CN"              },